C# was moved to Dev_Gen

+++ PulseAnalyzer/Source/dspPdw.h
@@ -23,9 +23,9 @@ public:
    // Members
 
    int    mSeqNum;
-   int    mToa;
+   double mToa;
    double mAmplitude;
-   int    mPulseWidth;
+   double mPulseWidth;
 
    //--------------------------------------------------------------------------
    // Constructors
